updates for hysteresis
