+++ src/m3util/viz/__init__.py
@@ -7,5 +7,6 @@ from . import printing
 from . import arrows
 from . import images
 from . import lines
+from . import positioning
 
-__all__ = ["movies", "layout", "style", "printing", "arrows", images, lines]
+__all__ = ["movies", "layout", "style", "printing", "arrows", "images", "lines", "positioning"]
